backend/requirements.txt
fastapi
uvicorn
python-dotenv
googlemaps
google-generativeai
requests
pydantic
httpx
google-api-python-client
google-auth-httplib2
google-auth-oauthlib